+++ ForestManagement/functions/exchangeschema/Test-FMExchangeSchema.ps1
@@ -1,6 +1,5 @@
-﻿function Test-FMExchangeSchema
-{
-<#
+﻿function Test-FMExchangeSchema {
+	<#
 	.SYNOPSIS
 		Tests, whether the desired Exchange version has already been applied to the Forest.
 	
@@ -27,8 +26,7 @@
 		$Credential
 	)
 	
-	begin
-	{
+	begin {
 		$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include Server, Credential
 		$parameters['Debug'] = $false
 		Assert-ADConnection @parameters -Cmdlet $PSCmdlet
@@ -36,8 +34,7 @@
 		Assert-Configuration -Type ExchangeSchema -Cmdlet $PSCmdlet
 		
 		#region Utility Functions
-		function Get-ExchangeRangeUpper
-		{
+		function Get-ExchangeRangeUpper {
 			[CmdletBinding()]
 			param (
 				[hashtable]
@@ -48,8 +45,7 @@
 			(Get-ADObject @parameters -SearchBase $rootDSE.schemaNamingContext -LDAPFilter "(name=ms-Exch-Schema-Version-Pt)" -Properties rangeUpper).rangeUpper
 		}
 		
-		function Get-ExchangeObjectVersion
-		{
+		function Get-ExchangeObjectVersion {
 			[CmdletBinding()]
 			param (
 				[hashtable]
@@ -60,8 +56,7 @@
 			(Get-ADObject @parameters -SearchBase $rootDSE.configurationNamingContext -LDAPFilter '(objectClass=msExchOrganizationContainer)' -Properties ObjectVersion).ObjectVersion
 		}
 		
-		function Get-ExchangeOrganizationName
-		{
+		function Get-ExchangeOrganizationName {
 			[CmdletBinding()]
 			param (
 				[hashtable]
@@ -73,46 +68,56 @@
 		}
 		#endregion Utility Functions
 	}
-	process
-	{
+	process {
 		$forest = Get-ADForest @parameters
 		$schemaVersion = Get-ExchangeRangeUpper -Parameters $parameters
 		$objectVersion = Get-ExchangeObjectVersion -Parameters $parameters
-		$displayName = (Get-ExchangeVersion | Where-Object RangeUpper -eq $schemaVersion | Where-Object ObjectVersionConfig -EQ $objectVersion | Sort-Object Name | Select-Object -Last 1).Name
-		
+		$displayName = (Get-AdcExchangeVersion | Where-Object RangeUpper -EQ $schemaVersion | Where-Object ObjectVersionConfig -EQ $objectVersion | Sort-Object Name | Select-Object -Last 1).Name
+		$splitPermissionsEnabled = Test-ADObject @parameters -Identity ("OU=Microsoft Exchange Protected Groups,%DomainDN%" | Resolve-String)
+
 		$adData = [pscustomobject]@{
-			SchemaVersion = $schemaVersion
-			ObjectVersion = $objectVersion
-			DisplayName   = $displayName
+			SchemaVersion    = $schemaVersion
+			ObjectVersion    = $objectVersion
+			DisplayName      = $displayName
 			OrganizationName = Get-ExchangeOrganizationName -Parameters $parameters
+			SplitPermissions = $splitPermissionsEnabled
 		}
 		Add-Member -InputObject $adData -MemberType ScriptMethod -Name ToString -Value {
 			if ($this.DisplayName) { $this.DisplayName }
 			else { '{0} : {1}' -f $this.SchemaVersion, $this.ObjectVersion }
 		} -Force
 		$configuredData = Get-FMExchangeSchema
-		
-		if ($configuredData.SchemaOnly)
-		{
-			if (-not $schemaVersion)
-			{
-				New-TestResult -ObjectType ExchangeSchema -Type CreateSchema -Identity $forest -Server $Server -Configuration $configuredData -ADObject $adData
-			}
-			elseif ($configuredData.RangeUpper -gt $schemaVersion)
-			{
-				New-TestResult -ObjectType ExchangeSchema -Type UpdateSchema -Identity $forest -Server $Server -Configuration $configuredData -ADObject $adData
-			}
+
+		$common = @{
+			ObjectType    = 'ExchangeSchema'
+			Identity      = $forest
+			Server        = $Server
+			Configuration = $configuredData
+			ADObject      = $adData
 		}
-		else
-		{
-			if (-not $schemaVersion -or -not $objectVersion)
-			{
-				New-TestResult -ObjectType ExchangeSchema -Type Create -Identity $forest -Server $Server -Configuration $configuredData -ADObject $adData
+		
+		if ($configuredData.SchemaOnly) {
+			if (-not $schemaVersion) {
+				New-TestResult @common -Type CreateSchema
 			}
-			elseif (($configuredData.RangeUpper -gt $schemaVersion) -or ($configuredData.ObjectVersion -gt $objectVersion))
-			{
-				New-TestResult -ObjectType ExchangeSchema -Type Update -Identity $forest -Server $Server -Configuration $configuredData -ADObject $adData
+			elseif ($configuredData.RangeUpper -gt $schemaVersion) {
+				New-TestResult @common -Type UpdateSchema
 			}
+			return
+		}
+		
+		if (-not $schemaVersion -or -not $objectVersion) {
+			New-TestResult @common -Type Create
+			return
+		}
+		if (($configuredData.RangeUpper -gt $schemaVersion) -or ($configuredData.ObjectVersion -gt $objectVersion)) {
+			New-TestResult @common -Type Update
+		}
+		if ($splitPermissionsEnabled -and -not $configuredData.SplitPermission) {
+			New-TestResult @common -Type DisableSplitP
+		}
+		if (-not $splitPermissionsEnabled -and $configuredData.SplitPermission) {
+			New-TestResult @common -Type EnableSplitP
 		}
 	}
 }