+++ ForestManagement/changelog.md
@@ -1,5 +1,10 @@
 ﻿# Changelog
 
+## 1.5.31 (2022-03-18)
+
+- Upd: ExchangeSchema - added support for Split Permission mode
+- Fix: ExchangeSchema - broken invocation
+
 ## 1.5.29 (2021-04-23)
 
 - New: Command Register-FMServer - define settings controlling server-site assignments.